use random walk prior in step wedge design
the model does not has divergent transitions, but max tree depth rate is 100%. It may because in the data generation, the slop of time spline (exponential distributed) change dramatically, but we force the spline coefficient to be similar in the stan code. Try1. remove the random walk prior and see if tree_max problem get solved. 2. change data generation process, the coefficient in spline does not change much